extract the database models into their own module
